chore: changing warehouse address name from plural to singluar
